Extend KDocs, extract ModalCommandConsequences in its own file
